Element buffer to use pixels instead of viewports
We measured the amount of scroll a user was able to perform in a single
swipe in both large and small screen devices. We determined that while
the amount of viewports scrolled changed, the absolute amount of pixels
was always about the same 7500px. So it makes more sense to use the
absolute measure than it does to use the viewport count.